requirements.txt
pykeepass
toml
python-daemon
pex
